Add early VisionSubsystem
Added a pretty simple VisionSubsystem for testing (simulation is kinda supported but it was killing my computer lol)

It did have a lot more implemented but I realized I was doing it in a pretty inefficient way so I had to remove it for now. Will add it back when it is better.

Signed-off-by: UHDbits <uhdbits@duck.com>