Update to WPILib 2024.3.1
Also add a constant that will disable vision, in case the vision doesn't work properly at any point. The way that the vision is disabled may need to be adjusted at some point depending on how we implement the vision into the robot, but for now this will be fine.

Signed-off-by: UHDbits <uhdbits@duck.com>