Add new speaker auto shoot command
This commit adds a brand new auto shoot command that is meant for the arm in a static position, and hopefully this will allow us to shoot a lot faster in both autonomous and teleop. This commit also tunes the swerve power limits and ramp rates, and creates the file for a new Center 4 Note auto.

Signed-off-by: Ashton A <uhdbits@duck.com>